Improved format-args security (-Wformat-literals, etc) (#1363)
- In Xcode, enabled -Wformat -Wformat-nonliteral -Wformat-security
- Fixed a couple of places that were accidentally passing non-literal
  strings to format args.
- Declared more functions as being __printflike to fix warnings.
- Updated Fleece, where I made similar changes.
- Add new format warnings to CMake project (Linux / mac)
    - Windows doesn't give useful information in this case due to system headers setting off the warning countless times

+++ Crypto/PublicKey.cc
@@ -67,7 +67,7 @@ namespace litecore { namespace crypto {
             case KeyFormat::Raw:
                 return publicKeyRawData();
             default:
-                Assert(false, "Invalid key format received (%d)", format);
+                Assert(false, "Invalid key format received (%d)", (int)format);
         }
     }
 
@@ -121,7 +121,7 @@ namespace litecore { namespace crypto {
             case KeyFormat::Raw:
                 return publicKeyRawData();
             default:
-                Assert(false, "Invalid key format received (%d)", format);
+                Assert(false, "Invalid key format received (%d)", (int)format);
         }
 
     }
