+++ LiteCore/Storage/SQLiteDataFile.cc
@@ -228,7 +228,7 @@ namespace litecore {
                       "  kvmeta (name TEXT PRIMARY KEY, lastSeq INTEGER DEFAULT 0, purgeCnt INTEGER DEFAULT 0) WITHOUT ROWID; "
                       "PRAGMA user_version=%d; "
                       "END;",
-					  SchemaVersion::Current));
+					  (int)SchemaVersion::Current));
                 Assert(intQuery("PRAGMA auto_vacuum") == 2, "Incremental vacuum was not enabled!");
                 _schemaVersion = SchemaVersion::Current;
                 // Create the default KeyStore's table:
@@ -308,7 +308,7 @@ namespace litecore {
                                        function_ref<void()> upgrade)
     {
         auto logUpgrade = [&](const char *msg) {
-            logInfo("SCHEMA UPGRADE (%d-%d) %-s", _schemaVersion, minVersion, msg);
+            logInfo("SCHEMA UPGRADE (%d-%d) %-s", (int)_schemaVersion, (int)minVersion, msg);
         };
 
         if (_schemaVersion < minVersion) {
@@ -327,7 +327,7 @@ namespace litecore {
                 _exec("BEGIN");
                 inTransaction = true;
                 upgrade();
-                _exec(format("PRAGMA user_version=%d; END", minVersion));
+                _exec(format("PRAGMA user_version=%d; END", (int)minVersion));
             } catch (const SQLite::Exception &x) {
                 // Recover if the db file itself is read-only (but not opened with writeable=false)
                 if (x.getErrorCode() == SQLITE_READONLY) {
